feat(forks): Add EIP-7928
