+++ packages/testing/src/execution_testing/forks/forks/eips/__init__.py
@@ -1,0 +1,1 @@
+"""Listings of all EIPs, current and upcoming."""
