+++ packages/testing/src/execution_testing/forks/forks/eips/eip_7928.py
@@ -1,0 +1,54 @@
+"""
+EIP-7928: Block-Level Access Lists.
+
+Enforced block access lists with state locations and post-transaction state
+diffs.
+
+https://eips.ethereum.org/EIPS/eip-7928
+"""
+
+from dataclasses import replace
+
+from ...gas_costs import GasCosts
+
+
+class EIP7928:
+    """EIP-7928 fork."""
+
+    @classmethod
+    def header_bal_hash_required(cls) -> bool:
+        """
+        Header must contain block access list hash (EIP-7928).
+        """
+        return True
+
+    @classmethod
+    def gas_costs(cls) -> GasCosts:
+        """
+        The cost per block access list item is introduced in EIP-7928.
+        """
+        return replace(
+            super(EIP7928, cls).gas_costs(),  # type: ignore[misc]
+            GAS_BLOCK_ACCESS_LIST_ITEM=2000,
+        )
+
+    @classmethod
+    def empty_block_bal_item_count(cls) -> int:
+        """
+        Return the BAL item count for an empty Amsterdam block.
+
+        Four system contracts produce 15 items:
+          EIP-4788 beacon roots:           1 address + 1 write + 1 read = 3
+          EIP-2935 history storage:        1 address + 1 write          = 2
+          EIP-7002 withdrawal requests:    1 address + 4 reads          = 5
+          EIP-7251 consolidation requests: 1 address + 4 reads          = 5
+        """
+        return 15
+
+    @classmethod
+    def engine_execution_payload_block_access_list(cls) -> bool:
+        """
+        From Amsterdam, engine execution payload includes `block_access_list`
+        as a parameter.
+        """
+        return True
